chore: final touches before publication
