+++ common/main/archetype-resources/.editorconfig
