+++ common/main/archetype-resources/.gitattributes
