+++ common/main/archetype-resources/.gitignore
