+++ java-advanced/src/test/resources/projects/run-extratests-it/reference/pom.xml
@@ -37,7 +37,7 @@
   <licenses>
     <license>
       <name>MIT License</name>
-      <url>http://www.opensource.org/licenses/mit-license.php</url>
+      <url>https://opensource.org/licenses/MIT</url>
       <distribution>repo</distribution>
     </license>
   </licenses>
