+++ java-advanced/src/test/resources/projects/run-extratests-it/verify.groovy
@@ -1,27 +1,10 @@
-import static java.nio.file.Files.isReadable
+evaluate(new File(basedir, "tools.groovy"))
 
-import java.util.Properties
-
-def basePath = basedir.toPath()
-def propsPath = basePath.resolve "archetype.properties"
-
-def archetypeProps = new Properties()
-propsPath.withInputStream {
-    archetypeProps.load(it)
-}
-
-def targetArtifactId = archetypeProps.get "artifactId"
-def targetPath = basePath.resolve "project/$targetArtifactId/target"
-
-def targets = [
+verifyTargetFiles([
     "junit-platform/console-launcher.out.log",
     "junit-platform/TEST-junit-jupiter.xml",
     "pit-reports/index.html",
     "pit-reports/linecoverage.xml",
     "site/jacoco/index.html",
     "site/jacoco/jacoco.xml",
-    "it/simple-it/build.log"]
-
-targets.every {
-    isReadable targetPath.resolve(it)
-}
+    "it/simple-it/build.log"])
