+++ java-full/pom.xml
@@ -13,13 +13,15 @@
     A plain Java project, includes unit tests, code coverage, mutation tests, integration tests,
     packaging, signing, licensing, and deployment configuration.
   </description>
-  <url>https://github.com/TomerFi/tomfi-archetypes</url>
-  <inceptionYear>2022</inceptionYear>
 
   <parent>
     <groupId>info.tomfi.archetypes</groupId>
     <artifactId>tomfi-archetypes-parent</artifactId>
     <version>0.1.0-SNAPSHOT</version>
   </parent>
+
+  <properties>
+    <main.basedir>${project.basedir}/..</main.basedir>
+  </properties>
 
 </project>
