+++ java-playground-mod/pom.xml
@@ -12,13 +12,15 @@
   <description>Java project for playaround, hacking, or just testing some stuff out.
     A modular (JPMS) Java project, includes unit tests with JUnit's plarform.
   </description>
-  <url>https://github.com/TomerFi/tomfi-archetypes</url>
-  <inceptionYear>2022</inceptionYear>
 
   <parent>
     <groupId>info.tomfi.archetypes</groupId>
     <artifactId>tomfi-archetypes-parent</artifactId>
     <version>0.1.0-SNAPSHOT</version>
   </parent>
+
+  <properties>
+    <main.basedir>${project.basedir}/..</main.basedir>
+  </properties>
 
 </project>
