--- java-playground-mod/src/main/resources/archetype-resources/.editorconfig
@@ -1,50 +1,0 @@
-# editorconfig.org
-root = true
-
-[*]
-charset = utf-8
-end_of_line = lf
-insert_final_newline = true
-
-[**.java]
-indent_style = space
-indent_size = 2
-max_line_length = 100
-end_of_line = lf
-trim_trailing_whitespace = true
-
-[**.{xml,fxml}]
-indent_style = space
-indent_size = 2
-end_of_line = lf
-max_line_length = 120
-trim_trailing_whitespace = true
-
-[**.{yaml,yml}]
-indent_style = space
-indent_size = 2
-end_of_line = lf
-max_line_length = 100
-trim_trailing_whitespace = true
-
-[**.md]
-trim_trailing_whitespace = false
-max_line_length = 120
-
-[**.json]
-indent_style = space
-indent_size = 2
-end_of_line = lf
-
-[**/hooks/**]
-end_of_line = lf
-
-[**.config]
-end_of_line = lf
-
-[**.properties]
-end_of_line = lf
-
-[Dockerfile*]
-indent_style = space
-indent_size = 4
