--- java-playground-mod/src/main/resources/archetype-resources/.gitattributes
@@ -1,131 +1,0 @@
-# Normalize text files (crlf -> lf)
-#* text=auto
-# Language specific
-*.py text eol=crlf diff=python
-*.pyc text eol=crlf diff=python
-*.cs text eol=crlf diff=csharp
-*.java text eol=lf diff=java
-*.groovy text eol=crlf diff=java
-*.proto text eol=crlf diff=text
-# Documents
-*.csv text eol=crlf
-*.doc text eol=crlf diff=text
-*.docx text eol=crlf diff=text
-*.pdf text eol=crlf diff=text
-*.rtf text eol=crlf diff=text
-# Documentation
-*.markdown text eol=lf
-*.md text eol=lf
-*.mdtxt text eol=lf
-*.mdtext text eol=lf
-*.txt text eol=lf
-*.rst text eol=lf
-CHANGELOG text eol=lf
-CONTRIBUTING text eol=lf
-*COPYRIGHT* text eol=lf
-LICENSE text eol=lf
-*README* text eol=lf
-# Configs
-*.cnf text eol=lf
-*.conf text eol=crlf
-*.config text eol=lf
-*.lock binary
-*.npmignore text eol=lf
-*.properties text eol=lf
-*.toml text eol=lf
-*.yaml text eol=lf
-*.yml text eol=lf
-.editorconfig text eol=lf
-.env text eol=lf
-package-lock.json binary
-Makefile text eol=lf
-makefile text eol=lf
-# Graphics
-*.bmp binary
-*.gif binary
-*.gifv binary
-*.jpg binary
-*.jpeg binary
-*.ico binary
-*.png binary
-*.svg text eol=lf
-*.svgz binary
-*.tif binary
-*.tiff binary
-*.wbmp binary
-*.webp binary
-# Scripts
-*.bash text eol=lf
-*.sh text eol=lf
-*.sql text eol=lf
-# Windows
-*.bat text eol=crlf
-*.cmd text eol=crlf
-*.ps1 text eol=crlf
-# Docker
-*.dockerignore text eol=lf
-*-dockerignore text eol=lf
-Dockerfile text eol=lf
-Dockerfile* text eol=lf
-# Git
-*.gitattributes text eol=lf
-.gitignore text eol=lf
-.gitmodules text eol=rf
-# Web files
-*.coffee text eol=lf
-*.css text eol=lf diff=css
-*.htm text eol=lf diff=html
-*.html text eol=lf diff=html
-*.ini text eol=crlf
-*.js text eol=lf
-*.json text eol=lf
-*.jsp text eol=lf
-*.jspf text eol=lf
-*.jspx text eol=lf
-*.jsx text eol=lf
-*.less text eol=lf
-*.ls text eol=lf
-*.map binary
-*.php text eol=lf diff=php
-*.scss text eol=lf diff=css
-*.xml text eol=lf
-*.xhtml text eol=lf diff=html
-# Binary files
-*.class binary
-*.ear binary
-*.so binary
-*.war binary
-*.db binary
-*.p binary
-*.pkl binary
-*.pickle binary
-*.pyc binary
-*.pyd binary
-*.pyo binary
-# Visual studio specific
-*.sln text eol=crlf
-*.csproj text eol=crlf
-*.vbproj text eol=crlf
-*.vcxproj text eol=crlf
-*.vcproj text eol=crlf
-*.dbproj text eol=crlf
-*.fsproj text eol=crlf
-*.lsproj text eol=crlf
-*.wixproj text eol=crlf
-*.modelproj text eol=crlf
-*.sqlproj text eol=crlf
-*.wmaproj text eol=crlf
-*.xproj text eol=crlf
-*.props text eol=crlf
-*.filters text eol=crlf
-*.vcxitems text eol=crlf
-
-# Git lfs
-*.exe filter=lfs diff=lfs merge=lfs -text
-*.jar filter=lfs diff=lfs merge=lfs -text
-*.jar binary
-*.dll filter=lfs diff=lfs merge=lfs -text
-*.7z filter=lfs diff=lfs merge=lfs -text
-*.gz filter=lfs diff=lfs merge=lfs -text
-*.tar filter=lfs diff=lfs merge=lfs -text
-*.zip filter=lfs diff=lfs merge=lfs -text
