+++ java-playground-mod/src/test/resources/projects/run-unittests-mod-it/reference/pom.xml
@@ -31,7 +31,7 @@
   <licenses>
     <license>
       <name>MIT License</name>
-      <url>http://www.opensource.org/licenses/mit-license.php</url>
+      <url>https://opensource.org/licenses/MIT</url>
       <distribution>repo</distribution>
     </license>
   </licenses>
