+++ java-playground-mod/src/test/resources/projects/run-unittests-mod-it/verify.groovy
@@ -1,22 +1,5 @@
-import static java.nio.file.Files.isReadable
+evaluate(new File(basedir, "tools.groovy"))
 
-import java.util.Properties
-
-def basePath = basedir.toPath()
-def propsPath = basePath.resolve "archetype.properties"
-
-def archetypeProps = new Properties()
-propsPath.withInputStream {
-    archetypeProps.load(it)
-}
-
-def targetArtifactId = archetypeProps.get "artifactId"
-def targetPath = basePath.resolve "project/$targetArtifactId/target"
-
-def targets = [
+verifyTargetFiles([
     "junit-platform/console-launcher.out.log",
-    "junit-platform/TEST-junit-jupiter.xml"]
-
-targets.every {
-    isReadable targetPath.resolve(it)
-}
+    "junit-platform/TEST-junit-jupiter.xml"])
